+++ packages/wiz-ui-next/src/components/base/inputs/search-input/search-input.vue
@@ -274,6 +274,11 @@ const props = defineProps({
     required: false,
     default: false,
   },
+  showParentLabel: {
+    type: Boolean,
+    required: false,
+    default: false,
+  },
 });
 
 const emit = defineEmits<{
@@ -282,12 +287,16 @@ const emit = defineEmits<{
   (e: "toggle", value: boolean): void;
 }>();
 
+const searchValue = ref("");
+
 const checkValues = computed({
   get: () => props.modelValue,
-  set: (value: number[]) => emit("update:modelValue", value),
+  set: (value: number[]) => {
+    emit("update:modelValue", value);
+    searchValue.value = "";
+  },
 });
 
-const searchValue = ref("");
 const filteredOptions = ref<SearchInputOption[]>([]);
 const selectedItem = ref<number[]>([]);
 const activeItem = ref<number | null>();
@@ -300,10 +309,17 @@ const valueToOption = computed(() => {
 
   const flatten = (options: SearchInputOption[]): SearchInputOption[] => {
     return options.flatMap((option) => {
-      if (option.children) {
-        return [option, ...flatten(option.children)];
-      }
-      return [option];
+      if (!option.children) return [option];
+
+      if (!props.showParentLabel) return [option, ...flatten(option.children)];
+
+      const children = option.children.map((child) => ({
+        ...child,
+        // 要件上、全角空白のため無視
+        // eslint-disable-next-line no-irregular-whitespace
+        label: `${option.label}　${child.label}`,
+      }));
+      return [option, ...flatten(children)];
     });
   };